git commit message generated by `debgpt --cmd 'git diff --staged' --ask 'briefly describe these changes. answer as short as possible' -Q`
addition of a new function called `command_line` in the `debian.py` file.
This function takes a command as input and returns the output of that command.
In the `main_cli.py` file, a new command-line argument `--cmd` is added,
which allows the user to include the command line output in the prompt.